Added goal achievement function to engine.py
This function takes one specific solution and evaluates it: checks if it satisfies all the requirements (like not exceeding the budget) and if it does, calculates the total score of this solution - the bigger, the better.

Additionally, formula in function_F was completely overhauled. Now it has sigmoid-like shape with respect to distance and allows to set maximum value.

Updated the engine.py so that all functions have exhaustive (at least in my opinion...) docstrings.

At the end of the script is a test block checking that everything works properly. This code will never be launched if this module is only imported, as it should.

TODO:
- now that goal achievement function is ready the only thing missing is some solution to the optimization problem of maximizing this function.