Modified symmetrize_numpy_matrix()
It now checks not only for total amount of dimensions, but also if the last two axes have the same size. If they do not, an error is thrown. This operation is impossible to perform with mismatching dimensions and would throw a numpy error anyway.

Also,fixed a bug that caused symmetrize_numpy_matrix() to fail when more than 2D were given.

Now it can work properly for any number of dimensions