Generalized goal_function_convenient()
Now it can take not only one 2D connection matrix, but also higher-dimension arrays.
When higher dimensional array is provided it is treated as a batch of 2D connection-matrices and instead of a scalar an ndarray is returned. It has the shape of given connetion-matrix but with the final two dimensions reduced.