Expanded the mutate_bool_ndarray() method
It now allows to provide another array with 'spare indexes' - a numpy ndarray specifying, which elements of `arr` will not be mutated. This is required to establish elitism.

Also, it now allows to specify if the method works on the original or creates a new copy. This is useful when the original is needed to remain intact