Implemented rewards matrix into the genetic algorithm.
It significantly improved its performance. The improvement is heavily dependent on the parameters. During testing for some cases got improvements of over an order of magnitude (18 times faster for my biggest 'simulation' - 17 seconds instead of over 5 min) and for other cases improvements were negligible. On average (I did purely arbitrary and random tests) it was around 5 times faster.
It seems like the more sparse the connections, the smaller the improvement.

Also leveraged the fact that rewards_matrix has to be symmetric and changed the method of calculating it so that only one half is computed and other half is just copied using symmetrize_numpy_matrix(). This effectively cut time required to create it in half (since the copying is done in microseconds and computations in milliseconds - I checked)

TODO:
- maybe try to lazily calculate the reward matrix (I do not know if it will help, but I'm down to try and see for myself)